Fix wrong cmd to gen private key with passphrase
To generate a private key with passphrase an encryption must be specified `-aes128`. Otherwise a key without passphrase will be created and no error will occur. The stupid thing is that the PHP code does not throw an error if you specify a passphrase for a private key without passphrase. Therefore no one notices the error.